add image optimization, add clean architecture (folders, barrels, archives, etc), edit styles, icons and content of page sections.
