+++ app/(admin)/head.tsx
@@ -1,9 +1,10 @@
+
 export default function Head() {
   return (
     <>
       <title></title>
-      <meta content="width=device-width, initial-scale=1" name="viewport" />
-      <link rel="icon" href="/favicon.ico" />
+      <meta content='width=device-width, initial-scale=1' name='viewport' />
+      <link rel='icon' href='/favicon.ico' />
     </>
   )
 }
