+++ app/head.tsx
@@ -1,9 +1,10 @@
-'use client'
+'use client';
 
 import { usePathname } from 'next/navigation';
 
 const Head = () => {
 
+  // eslint-disable-next-line react-hooks/rules-of-hooks
   const pathname = usePathname() as string;
 
   const titleFixedToShow = pathname.slice(1).charAt(0).toUpperCase() + pathname.slice(2);
