+++ components/content/index.ts
@@ -1,6 +1,9 @@
 export * from './AboutMe';
+export * from './Hero';
 export * from './CurrentFavTech';
 export * from './FavProjects';
+export * from './SearchAllProjects';
 export * from './ProcessWork';
 export * from './TimeLineExp';
 export * from './ContactMe';
+export * from './CurrentTimeLineExp';
