+++ components/ui/ButtonTheme.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { useEffect, useState } from 'react';
 import { useTheme } from 'next-themes';
