+++ components/ui/LinksMenuNav.tsx
@@ -1,0 +1,43 @@
+import Link from 'next/link';
+
+const LinksMenu = [
+  {
+    name: 'Home',
+    path: '/',
+    delay: '150ms'
+  },
+  {
+    name: 'About',
+    path: '/about',
+    delay: '175ms'
+  },
+  {
+    name: 'Projects',
+    path: '/projects',
+    delay: '200ms'
+  },
+  {
+    name: 'Blog',
+    path: '/blog',
+    delay: '225ms'
+  },
+]
+
+export const LinksMenuNav = () => {
+  return (
+    <>
+      {
+        LinksMenu.map(({ name, path, delay }) => (
+          <li
+            key={name}
+            className='border-gray-300 dark:border-gray-700 text-gray-900 dark:text-gray-100 text-sm font-semibold'
+            style={{ transitionDelay: delay }}>
+            <Link href={path} className='pb-4'>
+              {name}
+            </Link>
+          </li>
+        ))
+      }
+    </>
+  )
+}
