+++ components/ui/index.ts
@@ -1,7 +1,10 @@
+export * from './Header';
 export * from './ButtonTheme';
 export * from '../content/CardProject';
 export * from './ExternalLink';
 export * from './Footer';
-export * from './Header';
 export * from './MobileMenuNav';
 export * from './NavItem';
+export * from './CirclesColorBg';
+export * from './LinksMenuNav';
+export * from './LinksNav';
