+++ components/utils/AnimationContainer.tsx
@@ -1,4 +1,5 @@
-import React from 'react'
+'use client';
+
 import { motion } from 'framer-motion';
 import { AnimationContainerProps } from '@/types';
