+++ components/utils/index.ts
@@ -1,3 +1,5 @@
-export * from './AnimationContainer';
 export * from './SectionContainer';
-export * from './getNameSkills';
+export * from './AnimationContainer';
+export * from './TitleSectionPageContainer';
+export * from './ShowSkills';
+export * from './HeaderAnimation';
